modify cout

+++ src/ResultWriter/WaveFieldWriter.cpp
@@ -304,7 +304,7 @@ void seissol::writer::WaveFieldWriter::init(unsigned int numVars,
 		} else {
 			numLowVars = WaveFieldWriterExecutor::NUM_LOWVARIABLES + m_numIntegratedVariables;
 		}
-		logInfo(rank) << "Total number of extra variables " << numLowVars << " PSTRAIN " << pstrain == 0L;
+		logInfo(rank) << "Total number of extra variables " << numLowVars << " PSTRAIN " << static_cast<int>(pstrain == 0L);
 
 		for (unsigned int i = 1; i < numLowVars; i++)
 			addBuffer(0L, pLowMeshRefiner->getNumCells() * sizeof(double));
